Include integration tests.
