+++ peach-redis-server/src/test/resources/reference.conf
@@ -1,0 +1,18 @@
+
+
+default-server {
+  akka.extensions = ["akka.contrib.pattern.ClusterReceptionistExtension"]
+  akka.actor.provider = "akka.cluster.ClusterActorRefProvider"
+  akka.remote.log-remote-lifecycle-events = off
+  akka.cluster.auto-down-unreachable-after = 10s
+}
+
+
+com.proteus.peach.redis.manager.session: {
+  id: "redis",
+  addresses: ["localhost"],
+  port: 6379
+}
+
+
+
